Move `SetColorScheme` command
