+++ wcfsetup/install/files/lib/acp/form/UserEditForm.class.php
@@ -19,9 +19,8 @@ use wcf\system\exception\UserInputException;
 use wcf\system\html\upcast\HtmlUpcastProcessor;
 use wcf\system\message\embedded\object\MessageEmbeddedObjectManager;
 use wcf\system\moderation\queue\ModerationQueueManager;
-use wcf\system\option\user\UserOptionHandler;
 use wcf\system\style\StyleHandler;
-use wcf\system\user\command\SetColorScheme;
+use wcf\command\user\SetColorScheme;
 use wcf\system\user\multifactor\Setup;
 use wcf\system\WCF;
 use wcf\util\StringUtil;
