+++ wcfsetup/install/files/lib/command/user/SetColorScheme.class.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace wcf\system\user\command;
+namespace wcf\command\user;
 
 use wcf\data\user\User;
 use wcf\data\user\UserAction;
@@ -8,21 +8,17 @@ use wcf\data\user\UserAction;
 /**
  * Sets the preferred color scheme of a user.
  *
- * @author Alexander Ebert
- * @copyright 2001-2023 WoltLab GmbH
- * @license GNU Lesser General Public License <http://opensource.org/licenses/lgpl-license.php>
- * @since 6.0
+ * @author      Alexander Ebert
+ * @copyright   2001-2023 WoltLab GmbH
+ * @license     GNU Lesser General Public License <http://opensource.org/licenses/lgpl-license.php>
+ * @since       6.2
  */
 final class SetColorScheme
 {
-    private readonly User $user;
-    private readonly string $colorScheme;
-
-    public function __construct(User $user, string $colorScheme)
-    {
-        $this->user = $user;
-        $this->colorScheme = $colorScheme;
-    }
+    public function __construct(
+        private readonly User $user,
+        private readonly string $colorScheme
+    ) {}
 
     public function __invoke(): void
     {
